as per comments, amends file structure to separate routes, removes login and register buttons and creates single Authbutton in shared components
