+++ src/App.jsx
@@ -1,12 +1,11 @@
 import React from "react";
-import { BrowserRouter as Router, Route, Routes } from "react-router-dom";
+import { BrowserRouter as Router, Routes, Route } from "react-router-dom";
 import Navbar from "./components/Navigation/Navbar";
+import AuthRoutes from "./routes/AuthRoutes"; 
+import Home from "./pages/Home";
 import About from "./pages/About";
 import Contact from "./pages/Contact";
-import SignIn from "./pages/SignIn";
-import Register from "./pages/Register";
-import Home from "./pages/Home";
-import Dashboard from "./pages/Dashboard";
+import DashboardRoutes from "./routes/DashboardRoutes"; 
 
 function App() {
   return (
@@ -17,13 +16,12 @@ function App() {
         </nav>
         <div>
           <Routes>
-            <Route path="/" exact element={<Home />} />
+            <Route path="/" element={<Home />} />
             <Route path="/about" element={<About />} />
             <Route path="/contact" element={<Contact />} />
-            <Route path="/signin" element={<SignIn />} />
-            <Route path="/register" element={<Register />} />
-            <Route path="/dashboard" element={<Dashboard />} />
           </Routes>
+          <AuthRoutes />
+          <DashboardRoutes /> 
         </div>
       </div>
     </Router>
