+++ src/components/Profile/UserProfile.jsx
@@ -1,0 +1,12 @@
+import React from "react";
+
+const Contact = () => {
+
+    return(
+        <div>
+            <h2>User Profile Section</h2>
+        </div>
+    )
+}
+
+export default Contact
