--- src/shared/registerButton/RegisterButton.css
@@ -1,17 +1,0 @@
-.register-button {
-    display: inline-block;
-    padding: 10px 20px;
-    margin: 10px;
-    background-color: #007efd;
-    color: white;
-    border: none;
-    border-radius: 5px;
-    font-size: 16px;
-    cursor: pointer;
-    text-decoration: none;
-    transition: background-color 0.3s ease;
-  }
-  
-  .register-button:hover {
-    background-color: #1500ff;
-  }
