--- src/shared/signinButton/SigninButton.css
@@ -1,17 +1,0 @@
-.signin-button {
-    display: inline-block;
-    padding: 10px 20px;
-    margin: 10px;
-    background-color: #4CAF50;
-    color: white;
-    border: none;
-    border-radius: 5px;
-    font-size: 16px;
-    cursor: pointer;
-    text-decoration: none;
-    transition: background-color 0.3s ease;
-  }
-  
-  .signin-button:hover {
-    background-color: #45a049;
-  }
